Feedback app was implemented. All related things were moved to feedback app
