+++ apps/partner/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 from django.contrib.gis import admin as gis_admin
-from .models import Establishment, Feedback, FeedbackAnswer
+from .models import Establishment
 
 
 class EstablishmentAdmin(gis_admin.OSMGeoAdmin):
@@ -8,5 +8,3 @@ class EstablishmentAdmin(gis_admin.OSMGeoAdmin):
 
 
 admin.site.register(Establishment, EstablishmentAdmin)
-admin.site.register(Feedback)
-admin.site.register(FeedbackAnswer)
