+++ apps/partner/views.py
@@ -16,18 +16,13 @@ from happyhours.permissions import (
     IsAdmin,
     IsPartnerOwner,
     IsPartnerUser,
-    IsUserObjectOwner,
 )
 from .serializers import (
     EstablishmentSerializer,
     EstablishmentCreateUpdateSerializer,
     # MenuSerializer,
-    FeedbackCreateUpdateSerializer,
-    FeedbackSerializer,
-    FeedbackAnswerCreateUpdateSerializer,
-    FeedbackAnswerSerializer,
 )
-from .models import Establishment, Feedback, FeedbackAnswer
+from .models import Establishment
 from ..beverage.models import Beverage
 from ..beverage.serializers import BeverageSerializer
 
@@ -140,63 +135,3 @@ class MenuView(viewsets.ReadOnlyModelViewSet):
                 "No beverages found for this establishment or establishment does not exist."
             )
         return super().list(request, *args, **kwargs)
-
-
-@extend_schema(tags=["Feedbacks"])
-class FeedbackListView(ListAPIView):
-    serializer_class = FeedbackSerializer
-
-    def get_queryset(self):
-        establishment = self.request.resolver_match.kwargs['pk']
-        return Feedback.objects.filter(establishment=establishment)
-
-
-@extend_schema(tags=["Feedbacks"])
-class FeedbackCreateView(CreateAPIView):
-    queryset = Feedback.objects.all()
-    serializer_class = FeedbackCreateUpdateSerializer
-    permission_classes = [IsAuthenticated]
-
-
-class FeedbackPermissions:
-    def get_permissions(self):
-        if self.action == "retrieve":
-            permissions = [IsAuthenticated]
-        elif self.action in ("update", "partial_update", "destroy"):
-            permissions = [IsUserObjectOwner]
-        else:
-            permissions = [IsAdmin]
-        return [permission() for permission in permissions]
-
-
-@extend_schema(tags=["Feedbacks"])
-class FeedbackViewSet(
-    FeedbackPermissions, ViewSetMixin, RetrieveAPIView, UpdateAPIView,
-    DestroyAPIView
-):
-    queryset = Feedback.objects.all()
-
-    def get_serializer_class(self):
-        if self.action == "update":
-            return FeedbackAnswerCreateUpdateSerializer
-        return FeedbackSerializer
-
-
-@extend_schema(tags=["Feedbacks"])
-class FeedbackAnswerCreate(CreateAPIView):
-    queryset = FeedbackAnswer.objects.all()
-    serializer_class = FeedbackAnswerCreateUpdateSerializer
-    permission_classes = [IsAuthenticated]
-
-
-@extend_schema(tags=["Feedbacks"])
-class FeedbackAnswerViewSet(
-    FeedbackPermissions, ViewSetMixin, RetrieveAPIView, UpdateAPIView,
-    DestroyAPIView
-):
-    queryset = FeedbackAnswer.objects.all()
-
-    def get_serializer_class(self):
-        if self.action == "update":
-            return FeedbackAnswerCreateUpdateSerializer
-        return FeedbackAnswerSerializer
